register

+++ app/interface/passport-login/model/const.go
@@ -5,9 +5,9 @@ const (
 )
 
 const (
-	SourceWeb     = 1
-	SourceiOS     = 2
-	SourceAndroid = 3
+	SourceWeb     = int32(1)
+	SourceiOS     = int32(2)
+	SourceAndroid = int32(3)
 )
 
 const (
@@ -25,14 +25,14 @@ const (
 )
 
 const (
-	IdentityMobile = 1
-	IdentityEmail  = 2
+	IdentityMobile = int32(1)
+	IdentityEmail  = int32(2)
 )
 
 const (
-	ValcodeRegister       = 1
-	ValcodeForgetPassword = 2
-	ValcodeLogin          = 3
+	ValcodeRegister       = int32(1)
+	ValcodeForgetPassword = int32(2)
+	ValcodeLogin          = int32(3)
 )
 
 const (
